refactor: change repository functions to another file
Since there are too much url related functions, i moved them to another repository.
