+++ src/repositories/index.js
@@ -1,3 +1,4 @@
 export * as usersRepository from './usersRepository.js';
 export * as postsRepository from './postsRepository.js';
 export * as tagsRepository from './tagsRepository.js';
+export * as urlsRepository from './urlsRepository.js';
